+++ tests/test_shfqa_sweeper.py
@@ -33,8 +33,16 @@ def mock_TriggerConfig():
 
 
 @pytest.fixture()
-def sweeper_module(mock_connection, mock_shf_sweeper, session):
-    yield SHFQASweeper(mock_connection.return_value, session)
+def mock_sweeper_daq():
+    with patch(
+        "zhinst.toolkit.driver.modules.shfqa_sweeper.ziDAQServer", autospec=True
+    ) as connection:
+        yield connection
+
+
+@pytest.fixture()
+def sweeper_module(session, mock_sweeper_daq, mock_shf_sweeper):
+    yield SHFQASweeper(session)
 
 
 def test_repr(sweeper_module):
@@ -46,15 +54,17 @@ def test_missing_node(mock_connection, mock_shf_sweeper, session):
         "zhinst.toolkit.driver.modules.shfqa_sweeper.SweepConfig",
         make_dataclass("Y", fields=[("s", str, 0)], bases=(SweepConfig,)),
     ) as sweeper_config:
-        sweeper_module = SHFQASweeper(mock_connection.return_value, session)
+        sweeper_module = SHFQASweeper(session)
         assert sweeper_module.sweep.s() == 0
 
 
-def test_device(mock_connection, sweeper_module, session, mock_shf_sweeper):
+def test_device(
+    mock_connection, sweeper_module, session, mock_shf_sweeper, mock_sweeper_daq
+):
     assert sweeper_module.device() == ""
 
     sweeper_module.device("dev1234")
-    mock_shf_sweeper.assert_called_with(mock_connection.return_value, "dev1234")
+    mock_shf_sweeper.assert_called_with(mock_sweeper_daq(), "dev1234")
     assert sweeper_module.device() == "dev1234"
 
     connected_devices = "dev1234"
@@ -71,15 +81,17 @@ def test_device(mock_connection, sweeper_module, session, mock_shf_sweeper):
     assert sweeper_module.device() == session.devices["dev1234"]
 
 
-def test_update_settings(mock_connection, sweeper_module, mock_shf_sweeper):
+def test_update_settings(
+    mock_connection, sweeper_module, mock_shf_sweeper, mock_sweeper_daq
+):
     assert not sweeper_module.envelope.enable()
-    mock_shf_sweeper.assert_called_with(mock_connection.return_value, "")
+    mock_shf_sweeper.assert_called_with(sweeper_module._daq_server, "")
     # device needs to be set first
     with pytest.raises(RuntimeError) as e_info:
         sweeper_module._update_settings()
 
     sweeper_module.device("dev1234")
-    mock_shf_sweeper.assert_called_with(mock_connection.return_value, "dev1234")
+    mock_shf_sweeper.assert_called_with(mock_sweeper_daq(), "dev1234")
     sweeper_module._update_settings()
     mock_shf_sweeper.return_value.configure.assert_called_once()
     assert "sweep_config" in mock_shf_sweeper.return_value.configure.call_args[1]